feat: hydra reactive reactor - initial commit
