+++ hydra-java-client/pom.xml
@@ -43,14 +43,14 @@
         <dependency>
             <groupId>org.cardanofoundation</groupId>
             <artifactId>hydra-java-cardano-client-lib-adapter</artifactId>
-            <version>0.0.6</version>
+            <version>0.0.7-SNAPSHOT</version>
             <scope>test</scope>
         </dependency>
 
         <dependency>
             <groupId>org.cardanofoundation</groupId>
             <artifactId>hydra-java-test-containers-support</artifactId>
-            <version>0.0.6</version>
+            <version>0.0.7-SNAPSHOT</version>
             <scope>test</scope>
         </dependency>
