+++ hydra-java-reactive-reactor-client/src/main/java/org/cardanofoundation/hydra/reactor/TxResult.java
