+++ hydra-java-reactive-reactor-client/src/main/java/org/cardanofoundation/hydra/reactor/commands/AbortHeadCommand.java
@@ -1,0 +1,9 @@
+package org.cardanofoundation.hydra.reactor.commands;
+
+public class AbortHeadCommand {
+
+    public static String key() {
+        return AbortHeadCommand.class.getName();
+    }
+
+}
