+++ hydra-java-reactive-reactor-client/src/main/java/org/cardanofoundation/hydra/reactor/commands/CloseHeadCommand.java
@@ -1,0 +1,9 @@
+package org.cardanofoundation.hydra.reactor.commands;
+
+public class CloseHeadCommand {
+
+    public static String key() {
+        return CloseHeadCommand.class.getName();
+    }
+
+}
