+++ hydra-java-reactive-reactor-client/src/main/java/org/cardanofoundation/hydra/reactor/commands/CommittedCommand.java
@@ -1,0 +1,9 @@
+package org.cardanofoundation.hydra.reactor.commands;
+
+public class CommittedCommand {
+
+    public static String key() {
+        return CommittedCommand.class.getName();
+    }
+
+}
