+++ hydra-java-reactive-reactor-client/src/main/java/org/cardanofoundation/hydra/reactor/commands/FanOutHeadCommand.java
@@ -1,0 +1,9 @@
+package org.cardanofoundation.hydra.reactor.commands;
+
+public class FanOutHeadCommand {
+
+    public static String key() {
+        return FanOutHeadCommand.class.getName();
+    }
+
+}
