+++ hydra-java-reactive-reactor-client/src/main/java/org/cardanofoundation/hydra/reactor/commands/InitHeadCommand.java
@@ -1,0 +1,9 @@
+package org.cardanofoundation.hydra.reactor.commands;
+
+public class InitHeadCommand {
+
+    public static String key() {
+        return InitHeadCommand.class.getName();
+    }
+
+}
