+++ hydra-java-reactive-reactor-client/src/main/java/org/cardanofoundation/hydra/reactor/commands/ReadyToFanOutHeadCommand.java
@@ -1,0 +1,9 @@
+package org.cardanofoundation.hydra.reactor.commands;
+
+public class ReadyToFanOutHeadCommand {
+
+    public static String key() {
+        return ReadyToFanOutHeadCommand.class.getName();
+    }
+
+}
